refactor: refactor components
