+++ package.json
@@ -37,6 +37,8 @@
     "@radix-ui/react-toast": "^1.1.4",
     "@radix-ui/react-toggle": "^1.0.3",
     "@radix-ui/react-tooltip": "^1.0.6",
+    "@starknet-react/chains": "^0.1.7",
+    "@starknet-react/core": "^2.8.0",
     "@vercel/analytics": "^1.0.2",
     "axios": "^1.5.0",
     "class-variance-authority": "^0.7.0",
@@ -44,6 +46,7 @@
     "cmdk": "^0.2.0",
     "date-fns": "^2.30.0",
     "dayjs": "^1.11.9",
+    "get-starknet-core": "^3.3.0",
     "lodash-es": "^4.17.21",
     "lucide-react": "^0.276.0",
     "next": "^14.2.3",
@@ -52,15 +55,17 @@
     "react-day-picker": "^8.8.1",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.46.1",
+    "react-hot-toast": "^2.4.1",
     "react-router-dom": "^6.15.0",
     "react-use": "^17.4.0",
+    "starknet": "^5.24.3",
     "swr": "^2.2.2",
     "tailwind-merge": "^1.14.0",
     "tailwind-variants": "^0.1.18",
     "tailwindcss-animate": "^1.0.7",
     "wasm-cairo": "^0.8.8",
     "zod": "^3.22.2",
-    "zustand": "^4.4.1"
+    "zustand": "^4.4.3"
   },
   "devDependencies": {
     "@types/lodash-es": "^4.17.9",
